requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pandas==2.3.2
openpyxl==3.1.5
reportlab==4.4.3
SQLAlchemy==2.0.39
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
argon2-cffi==23.1.0
bcrypt==4.0.1
SQLAlchemy==2.0.39
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
